added delete warnings, fixed userprofiles css

+++ src/components/books/bookShow.js
@@ -33,10 +33,12 @@ class BookShow extends React.Component {
   handleDelete(e) {
     e.preventDefault()
     console.log(`Bearer ${Auth.getToken()}`)
-    axios.delete(`/api/books/${this.props.match.params.id}`,
-      { headers: { Authorization: `Bearer ${Auth.getToken()}`}})
-      .then(() => this.props.history.push('/books/'))
-      .catch(err => this.setState({errors: err.response.data.errors}))
+    if (window.confirm('Delete the item?')) {
+      axios.delete(`/api/books/${this.props.match.params.id}`,
+        { headers: { Authorization: `Bearer ${Auth.getToken()}`}})
+        .then(() => this.props.history.push('/books/'))
+        .catch(err => this.setState({errors: err.response.data.errors}))
+    }
   }
